+++ facade/src/main/scala/react/semanticui/elements/container/package.scala
@@ -1,0 +1,5 @@
+package react.semanticui.elements
+
+package container {}
+
+package object container {}