+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.0.9
 
+* Add Container
 * Add Loader
 * Upgrade to react-semantic-ui 0.86.0